# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Project",
        "description": "Demo my Project ",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "d9261f9d7c0fb3ad438f876fba448ef4",
                "responses": {
                    "200": {
                        "description": "List Products."
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product with the provided data.",
                "operationId": "4da818c59d83cad86b9b82ea8e156372",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "400": {
                        "description": "Bad request. Invalid or missing parameters"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "e4a411b37ef1e3380ce5b77a508eb38f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id to get",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One category found."
                    },
                    "404": {
                        "description": "Category not found"
                    },
                    "400": {
                        "description": "Request validation fault"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "6c0305fa15c96d96eb3d04b299a14d49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id to edit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "double"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Validation has fault"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a Product",
                "description": "Deletes a specific Product based on the provided ID.",
                "operationId": "0d9d4ee54e991a67efac851847066b97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}